requirements.txt
Flask==3.0.0
python-dateutil==2.8.2
Werkzeug==3.0.1
geoip2==4.7.0
requests==2.31.0
